requirements.txt
# Telegram Bot (совместимо с supabase 2.17.0+)
python-telegram-bot==20.8

# AI & ML
google-generativeai>=0.3.2
openai-whisper>=20231117

# Database (версия с поддержкой httpx 0.26+)
supabase==2.17.0

# Utilities
python-dotenv>=1.0.0
pydub>=0.25.1
PyPDF2>=3.0.1
Pillow>=10.1.0
aiohttp>=3.9.1
ffmpeg-python>=0.2.0

